Moved CloseCode enum to own class and updated documentation
